JWTLogin/Backend/requirements.txt
asgiref
Django
django-cors-headers
django-environ
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
tzdata

# Successfully installed Django-5.0.6 PyJWT-2.8.0 asgiref-3.8.1 django-cors-headers-4.4.0 django-environ-0.11.2 djangorestframework-3.15.2 djangorestframework-simplejwt-5.3.1 pytz-2024.1 sqlparse-0.5.0 typing-extensions-4.12.2 tzdata-2024.1
